+++ src/qt_models/pid_tree_model.py
@@ -1,6 +1,6 @@
 from PyQt5.QtCore import Qt, QAbstractItemModel, QVariant, QModelIndex
 
-from src.qt_dialogs.tree_dialog.pidNode import PidNode
+from src.qt_models.tree_model_node import PidNode
 
 
 class PidTreeModel(QAbstractItemModel):
