+++ src/qt_models/tree_model_node.py
